[RENAME]: TourTheme 패키지 경로 변경
